fix: readiness/URL-validation bugs, rebuild portal, fix pre-existing IDF build errors
- local_ai_service: Authenticate() erforderte nur HTTP 2xx fuer
  success=true, ignorierte ob das konfigurierte Modell tatsaechlich in
  /v1/models gelistet ist. Jetzt success=model_listed, mit eigenem
  Statustext bei falschem/fehlendem Modell.
- local_ai_service: transcribe_url hatte keine Leer-Pruefung nach
  ApplySettingsPatch, anders als base_url. Jetzt symmetrisch validiert.
- Kconfig/sdkconfig.defaults: Platzhalter-Port 8010 fuer
  CONFIG_FOLLOWUP_LOCAL_AI_TRANSCRIBE_URL auf den echten, laufenden
  imkopfhaben-brain-Port 8000 korrigiert.
- Portal-Frontend-Bundle (index.html/index.js) mit npm run build neu
  erzeugt -- vorher zeigte der ausgelieferte Build-Artefakt noch
  "Gemini API Key", obwohl die TS-Quelle schon umgebaut war.
- Zwei vorbestehende, von den lokalen KI-Aenderungen unabhaengige
  Compile-Fehler unter ESP-IDF v5.4 gefixt (erstmals hier ueberhaupt
  kompiliert): xpowers_axp2101_driver.cc log_d()-Format-String-Mismatch
  (%x vs uint32_t, reiner Debug-Log, keine Verhaltensaenderung) und
  board_es8311_codec.cc .bclk_div-Feld, das laut eigenem Code-Kommentar
  im Master-Modus ohnehin wirkungslos war und in dieser IDF-Version aus
  i2s_std_clk_config_t entfernt wurde.

Voller Build (idf.py build, ESP-IDF v5.4, Target esp32s3) laeuft jetzt
durch: folloup_sticky.bin erzeugt, 65% Partitionsplatz frei.

+++ components/axp2101/xpowers_axp2101_driver.cc
@@ -2501,7 +2501,7 @@ bool Axp2101Driver::setInterruptImpl(uint32_t opts, bool enable)
     {
         int res = 0;
         uint8_t data = 0, value = 0;
-        log_d("%s - HEX:0x%x \n", enable ? "ENABLE" : "DISABLE", opts);
+        log_d("%s - HEX:0x%x \n", enable ? "ENABLE" : "DISABLE", static_cast<unsigned int>(opts));
         if (opts & 0x0000FF) {
             value = opts & 0xFF;
             // log_d("Write INT0: %x\n", value);
